+++ Challenges/21 - 30/(25)-1000-Digit-Fibonacci-Number.cs
@@ -4,6 +4,7 @@ using Enumerators;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Numerics;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -15,7 +16,7 @@ namespace Challenges
         //first index of the item in the fibb sequence to contain 1000 DIGITS
 
         //have a function the tretuns the n# of digits of an inputted nuber
-        public long RunChallenge()
+        public BigInteger RunChallenge()
         {
             var (number, index) = Fibonacci.Sequence().Select((number, i) => (number, i)).First(it => InputNumberAs.CountOfDigits(it.number) == 1000);
             var bar = index;
