+++ Challenges/21 - 30/(26)-Reciprocal-Cycles.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Numerics;
 using Enumerators;
 using ProjectEuler;
 
@@ -8,7 +9,7 @@ namespace Challenges
 {
     public class Challenge26 : IRunChallenge
     {
-        public long RunChallenge()
+        public BigInteger RunChallenge()
         {
             return UnitFractions.LongestRecurringCycle(1000).Item2;
         }
