+++ Challenges/21 - 30/(28)-Number-Spiral-Diagonals.cs
@@ -1,13 +1,14 @@
 ﻿using System.Linq;
+using System.Numerics;
+using Enumerators;
 using ProjectEuler;
-using ProjectEuler.Enumerators;
 
 namespace Challenges
 {
     public class Challenge28 : IRunChallenge
     {
         public long sizeofGrid;
-        public long RunChallenge()
+        public BigInteger RunChallenge()
         {
             return SpiralNumberCorners.Sequence().TakeWhile(x => x <= sizeofGrid * sizeofGrid).Sum();
         }
