+++ Enumerators/PrimeNumbers.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Collections;
 using System.Linq;
+using System.Numerics;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -9,9 +10,9 @@ namespace Enumerators
 {
     public class PrimeNumbers 
     {
-        public static IEnumerable<long> Sequence()
+        public static IEnumerable<BigInteger> Sequence()
         {
-            long number = 1;
+            BigInteger number = 1;
             do
             {
                 if (CheckForPrime(number))
@@ -24,14 +25,14 @@ namespace Enumerators
             } while (true);
         }
 
-        public static bool CheckForPrime(long naturalNumber)
+        public static bool CheckForPrime(BigInteger naturalNumber)
         {
 
             if (naturalNumber <= 1) return false;
             if (naturalNumber == 2) return true;
             if (naturalNumber % 2 == 0) return false;
 
-            var boundary = (long)Math.Floor(Math.Sqrt(naturalNumber));
+            var boundary = (long)Math.Floor(Math.Sqrt((double)naturalNumber));
 
             for (long i = 3; i <= boundary; i += 2)
                 if (naturalNumber % i == 0)
@@ -39,9 +40,5 @@ namespace Enumerators
 
             return true;
         }
-
     }
-
-
-    
 }
